pyproject.toml
[tool.poetry]
name = "project-data-science-team2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "project_data_science_team2"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.1"
pandas = "^2.0.3"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pillow = "^10.0.0"
pytest = "^7.4.0"
sphinx = "^7.1.1"
httpx = "^0.24.1"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
